pyproject.toml
[tool.poetry]
name = "zbots-python-lib"
version = "0.1.0"
description = ""
authors = ["Zbot <>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
"aiohttp" = "^2.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"